+++ CHANGELOG.md
@@ -7,6 +7,21 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Feature: Add `variant` prop to `RadioButton`, with default value `andromeda`.
 - Feature: Add `design` prop to `RadioButton`, with default value `disc`.
 - Feature: Better accessibility on `RadioButtonSet`.
+- Feature: Add `variant` prop to `TextInput` with `andromeda` as default value.
+
+## [2.98.0] - 2020-11-23
+
+Features:
+- Add `variant` prop to `VerticalStepper`, with default value `andromeda`.
+- Add `DotIcon` component.
+
+Fix:
+- `ImageCropper` `onChange` firing on uploaded images.
+
+## [2.97.0] - 2020-11-16
+
+Feature:
+- Add `CreditCardForm` component.
 
 ## [2.96.0] - 2020-11-12